chore: Update GTK UI properties for scroll view
This commit updates the properties of the GTK UI scroll view in the `ui.glade` file. The `vscrollbar-policy` property is changed from "always" to "automatic" and the `expand` property is set to true. These changes improve the user experience by allowing the vertical scrollbar to appear only when necessary and ensuring the scroll view expands to fill available space.