chore: Update dependencies in Cargo.toml
The code changes in the `Cargo.toml` file update the dependencies to their latest stable versions. This includes updating `gio` to version `0.9.1`, `comrak` to version `0.16`, `horrorshow` to version `0.8.4`, `reqwest` to version `0.12.5`, `serde_json` to version `1.0.121`, `tokio` to version `1.39.2`, `anyhow` to version `1.0.86`, and `glib` to version `0.20.0`. These updates ensure that the project is using the latest versions of the dependencies, which may include bug fixes, performance improvements, and new features.